ChangeListener: Always use lazyPush
Several of these listeners were using a plain push, which immediately
enqueues the job in the middle of execution. This opens up the
possibility that jobs could start running before the replicas have
seen the initial change, or even before the initial change is commited
to the database.

Switch to lazyPush which uses DeferredUpdates to delay submitting the
jobs until the request is in the shutdown phase. This should help
limit, but not remove, the potential for race conditions with the jobs
reading what was writen by the initial request.

Bug: T342562
Change-Id: I16b6410df4c718742bb4c99f213fb84135ef666e