Adjust Searcher::get to recognize response errors
As we've found in the past, Elastica's error handling does not
recognize errors that come from intermediaries. When this method
sees a 5xx error but without a json response in the expected format,
such as if a proxy like envoy has a problem, it passes through with
an empty response array.

Update to handle this like we have in other places, recognizing the
error and transforming it into a ResponseException. In testing with
real errors the ElasticaErrorHandler correctly extracts the error
message from envoy in these failed responses and logs an appropriate
failure.

Change-Id: I4002532054c5ae6fa50ddf26e8db87ff3bbd9155