Run CheckerJob against read-only clusters
Adjustments were made to JobTraits::decideClusters so it could source
the read only clusters for CheckerJob, but it seems enabling this
functionality was lost in a refactor.  Pass the read-only flag to
decideClusters so this starts checking the newly read-only wikis.

While here add a slight bit of context to the error messages when
a checker is removed from the group.

Bug: T354793
Change-Id: Ia4f1d2a67b1f2f97157ef97d63890115f824a379