Change CirrusSearchClusters merge strategy to provide_default
The merge strategy for an extension defaults to array_merge, which
means the default cluster defined here will be merged with any user
provided configuration. In the case of defining clusters to talk
to this doesn't make sense, user provided clusters should replace
any default cluster. As such change the merge strategy to
provide_default which avoids merging.

Bug: T354793
Change-Id: I02ec4a7f09b6cfb29b2a00931f4f60507e37f5a9