Drop CirrusSearchAllFields
this option has been "on" by default for a long time, long enough that
it is very possible that many assumptions are done regarding the
existence of these "all" fields.
Prefer to drop the possibility of having it disabled so that new
features can safely assume that these fields exist.

Change-Id: I6fe25d494d0ea4f50abfeac5809782354b95f78f