Add NearMatchFieldQueryBuilder
A simple query builder that takes either an already sanitized plain
string or a ParsedQuery and attempt to generate a multimatch query
against the near match field.
Purpose is to make it easier for query builders that are not based on
the old parser to generate a nearmatch clause that ressembles what we
generate from the FullTextQueryStringQueryBuilder.
A small refactoring was done to re-use the query building part.

Bug: T353683
Change-Id: I80cb42bb9bdf905407f75ea9ff870d127f8a5eaf