Connection: Correct read-only detection
The read only detection was failing on the edge case where the
Connection was setting it's read-only flag and the cluster was the
first in the list of read-only clusters. In that case array_search was
returning 0, and that was being cast into a false. Update the
comparison to only set read-only when array_search returns false, and
add a related test case.

Bug: T354793
Bug: T356526
Change-Id: I3d6282f6bee1e1fcd7228fdd7f3d4a49bd5c0468