Expose sanity checker over mw api
To enable the saneitizer to continue working with the streaming based
updater we are exposing the main checking process over the mw api.
This is marked internal, it is not generally intended for end users
although nothing will be harmed if they invoke it.

The overall plan is that the main loop and remediation moves over
to the streaming updater, while the validation of what belongs where
remains in CirrusSearch.

Bug: T358599
Change-Id: I1fed5266fe67b401a36f85da06e6d5c8a80a57f4