Group updates and decide clusters based on groups
We recently added the read only connection marker, but this doesn't
seem to have captured our use case particularly well. The goal was to
provide more granular control over what updates go where. Really we
have a few needs:

* We would like to disable page index updates while leaving
 archive and completion index updates alone.
* Index creation, mapping updates, etc all still need to be performed
* There is a related use case from saneitizer for the read only
 marker, to differentiate between a list of clusters to check, and a
 list of clusters to fix.

These all have the shared requirement of needing a list of clusters to
use in a particular context. This patch introduces an extension to
$wgCirrusSearchWriteClusters that makes separate groups of clusters
configurable, along with updating the places we do cluster selection
to reference these cluster lists. This isn't quite an exact match
for saneitizer's list of clusters to check, but close enough
that it should be ok.

$wgCirrusSearchWriteClusters now allows the value to be either a list
or an associative array. The list retains current functionality, an
associative array allows providing sets of clusters for the default
and specific use cases. These use cases are captured as UpdateKind::*
constants. Only the 'default' entry is required, all others are
optional and default to that entry.

Everywhere that references the existing CirrusSearchWriteClusters has
been updated to pass through an UpdateGroup::* constant. These
constants identify common use cases and allow us to specify the
clusters to use for a particular use case.

Change-Id: I337eef05bdf49cc40b808651169d0d932e57e57d