builddoc api: Check revision deletion when loading revisions
We've had an increase in error messages recently, it looks to be
because the new update pipeline is querying revisions a bit later than
the historical pipeline, running into more deleted revisions. The
deleted revisions fail deep in the document building, and previously
would result in uncaught exception logs.

Perhaps would should consider a deeper refactor, but for now pre-check
that the requested revisions are publicly viewable. If not respond
with an appropriate error message instead of blowing up. This should
also allow the caller to skip retrys, as we clearly indicate this
request will continue to fail.

Bug: T357066
Change-Id: Id64b80126f30b7328198b6dbd1072eff398de8e2