Disable Hiragana-to-Katakana Mapping
We had planned to expand `kana_map` from English to most other
languages, but testing revealed that doing the mapping
pre-tokenization interfered with the new ICU tokenizer's ability
to parse Japanese.

Enabling the mapping *after* tokenization didn't do much because
there aren't that many cross-kana terms in the dictionary used by
the ICU tokenizer.

In the spirit of cross-language analysis harmonization—and
because it's now clear that it's interfering with the ICU
tokenizer—remove the cross-kana mapping from the English analysis
chain.

Bug: T180387
Change-Id: I00e0e70160ee20e926e06027d301cd4a285431d1