Do not schedule LinksUpdate jobs if no clusters are writable
It is not necessary to flood the job queue if none of the clusters
are writable. Prior to emitting the LinkUpdate update job check if there
is at least one writable clusters and stop early if not.

Archive jobs have been moved to their own listener so that they can
continue being emitted regardless of the list of writable clusters for
UpdateGroup.PAGE.

Change-Id: Ic585aefe76373d59a0f5d930544f4db1549b350b