Add UpdateGroup for weighted tags
The weighted tags received in php have no equivalent elsewhere.
There is intent to develop a solution that involves pushing these
to eventgate, but that needs to be designed and communicated.

For now, define a custom update group for weighted tags so we can
point it at the normal clusters. Without this these updates would
be lost.

Change-Id: I860b6ab054ff82c874cd1d0e6080377abf53d719