Only reindex if settings or mappings have actually changed
In the past we expected an operator to recognize that half of a
reindex operation was a success, and to run the correct indexes
in a second operation that skips the already succesfull one.

But thats error prone, and tedious. Instead make it safe and fast
to run the index creation script even when there are no changes
to be applied. Accomplish this by creating the new index and
comparing it to the old index. If the settings and the mappings
are the same declare that there is nothing to do, delete the
new index, and continue on.

Change-Id: If67803c866dbb3b29271abeda8a510053a03956d