doc_values is already default true on long fields
We were setting doc_values to true here, but because that is
already the default value it wasn't coming back to us when
comparing index settings. The end result was that during index
creation the MappingValidator thought the page_id field needed an
update and was re-applying the mappings that are already set.

This should have already been a no-op on the elastic side, but it
will be slightly faster and a little cleaner to do it right.

Also adds a few default values to the searcher fixture testing,
matching the defaults in extension.json. For CI this should
make no differene, but on a dev instance that has
tests/jenkins/FullyFeaturedConfig.php included this lets the
tests work.

Change-Id: I5c8187473f1cada15aa9d5a7bd774c80564c8e23