Allow overriding the new index has changed reindexing check
It may be that an operator will decide they know better than this
check, perhaps we will need to ensure prior to an upgrade that
all indices were created with elasticsearch 7.10 or some such.
Add an option that lets the operator make the call.

Additionally, instead of exiting with success, give a specific
exit code when we refuse to reindex due to there being no change
to the index. This gives the reindex orchestration a chance to
know that the reindexing is completed even though no replacement
index was promoted.

Change-Id: Ifa69187f338b245d3628c14873c9ac1a4b07ee8b