Use upstream createNoOpMock in many more PHPUnit tests
I believe this makes the test quite a bit more predictable and
easier to read. It's much more obvious what code is supposed to be
called and what is not.

There is probably more, but I believe this is a good starting
point. I focussed on places where a mock was created but never any
expected method calls appended to the mock. Such a mock returns
null everywhere. This might or might not be intended.

Change-Id: Ie6a11943793797dd6a0b362817ec8ab9b845efc4