SiteMatrixInterwikiResolver should use the WAN cache
Refactor the code with proper injection, the code was particularly messy
for no good reasons.

The introduction of the WAN cache was initially done in
I6fea2941ffe696dd36e77ea49030f7f0c60ea054 but I think it was partially
done I believe because the code was particularly messy.

Change-Id: Ib33c52ec2886e2419133371d5f3c5f8d41fd6db4