Record a sampling of cache keys for page properties
We are having problems with an exceptionally low hit rate here.
We should be invoking this the same way from three different
requests, resulting in a hit rate of ~66%. But the observed hit
rate has declined to > 0.1%. It's not obvious what is going on,
but one possibility is that the cache key is varying in
unexpected ways, perhaps due to WikiPage::getTouched changing.
Log the cache key for a 1:1000 sample of pages so we can compare
them.

Change-Id: I25bbe276510c39fe5ad177f2800a7bc281a382db