browser tests: process redirects as a separate batch
We seem to fail quite frequently on redirects not being indexed.
Attempt to index them after we are sure that their targets are in the db
and the search index but also don't attempt to batch redirect that
targets the same page. There seems to be cases where the redirects array
is not properly updated with all existing redirects, reason is quite
unclear...

Change-Id: I8cbc215c43cb4cedaa7d4fa6277b589ecd9ffd99