browser tests: Reduce itermittent failures
This test, moving a page and leaving a redirect, is itermittently
failing in the CI pipeline. On review it looks like what is
happening is that there is no atomicity between the delete in the
old index and the insert to the new index. In some cases the
waitForOperation is seeing the redirect show up in the
destination index and continuing with the test, while the old
page has yet to be deleted. The test then fails by finding the
not-yet deleted page.

The provided solution is not amazing, it simply sleeps after the
waitForOperation and before performing the search, but it seems
to work. This was run in a loop for ~12 hours with the test
running every couple minutes without fail. Even if it's not
completely reliable, this should have significantly fewer
failures than in the past.

Change-Id: Ief07bb981f0ac132610a9dfbb081a4bf4d903bae