Re-add CirrusSearch prefix to statsd metrics
The statsd metrics in wmf.13 started recording in unexpected
locations. They were missing the CirrusSearch prefix to the
metrics. It looks like this was removed expecting
StatsFactory::withComponent to put it in place, and be formatted
by StatsdFormatter with that component, but the component here
doesn't make it down to that level.  Instead the only prefix that
appears to be prepended is 'MediaWiki.'.

Checked all invocations of StatsFactory::copyToStatsd. Tested by
configuring $wgStatsdServer and recording the sent metrics with
netcat.

Bug: T359033
Change-Id: I1d1968852052fad5cb986fe266ae72d79c7a83fb