Check the output of RevisionStore::getRevisionById
Do not take for granted the fact that ApiPageSet is flagging the
existing/missing revisions in ApiPageSet::initFromRevIDs. For some
reasons (lilkely because different db replica might be used)
RevisionStore::getRevisionById might still return null for a revision
marked as valid by ApiPageSet, double check the output
RevisionStore::getRevisionById and set the badrevids array accordingly
if the revision is not found by this service.

Bug: T370770
Change-Id: Ic77ac75aa46495656c7ddc257f7b1a3fead297fa