check_indices.py: Use parallelism
Collecting the configuration from mediawiki takes quite some
time. It's not that each individual invocation takes so long, but
we need to invoke the same script for each of 1000 wikis. This
was previously addressed by adding a cache of the mwscript
output, but that still makes the first run slow.

Updated script uses the same technique the reindexing script
does, run mwscript in parallel. Without parallism this takes more
then three minutes to populate the cache on first run, and then
~5s to fetch elasticsearch state and compare it against the
cache. The updated implementation now takes only ~35s on first
run with default parallelism of 8, ~22s if using a parallelism of
24.

With the runtime decreased from >3 minutes to ~35s it seems
reasonable to simply drop the run cache. Having the cache means
needing to think about how exactly to run the script, and having
something to cleanup after we are done. Removing the cache file
should simplify usage to a plain invocation of the script.

Change-Id: I371b71cf28253b9a277d1e4deea42269c2feac35