sanity: Handle redirect in wrong index
The saneitizer was correctly identifying when a redirect was
indexed, even if it was in an unexpected index. Unfortunately it
wasn't passing along which index it was found in so the
remediation applied was to delete it from the expected index.

Update the Remediator interface to pass along what index redirect
was found in, and update remediators to handle it appropriately.

Bug: T331127
Change-Id: Ia5c9031add91cdb51f6b94cea006c6f2f9b3ae59