Count matches from bad redirects
We've had some problems recently where bad redirects got indexed
into the search engine. It seems reasonable that they should be
dropped from the search results in the same way that we drop
results if the page title does not exist.

The redirects are slightly more complicated though, we don't have
a strong signal that the result was only a match due to the
redirect. As such clear the redirect from the search hit, but
keep the result in the results list.

While here introduce a new counter that tracks how often we drop
a result from the search results due to having a missing
revision.

Bug: T372446
Change-Id: I039681bbc0d3d78df95c761de4070f93c5a1348d