Bare bones OpenSearch 1.3.19 support
Changes the version check to expect opensearch 1.3, and updates
the readme to reference opensearch instead of elasticsearch. Many
references remain elsewhere.

When the connection is configured with a user/password and ssl
certs are appropriately handled this manages to index and search
for data while still using the old Elastica client libraries.

Bug: T372769
Change-Id: I1cf009ea0fecfe7ad2bee89598f8895a7d7134e3