Exclude connections from cirrus-config-dump
OpenSearch uses authentication by default, and our config dumping
api exposes that publicly. We couldn't think of a strong use case
for having this connection info generally available, so strip it
from the config dump output. If we find a use case the
username/password could be stripped, but it's questionable how
useful the connection info is without it.

Bug: T372769
Change-Id: I195100949d636e0c3ca43c75f36f612309c52a16