Exclude connections from cirrus-config-dump
OpenSearch uses authentication by default, and our config dumping
api exposes that publicly. This would be a similar concern for
elasticsearch+xpack. We couldn't think of a strong use case for
having this connection info generally available, so strip it from
the config dump output. If we find a use case the username /
password could be stripped, but it's questionable how useful the
connection info is without it.

Bug: T372769
Change-Id: Ia087612b8cd30b47127f2e14b8fcb1e0ff22111f