Bare bones OpenSearch 1.3.x support
Changes the version check to also accept opensearch 1.3 and
updates the readme to reflect this capability. Additionally
updates the experimental-highlighter check to accept the new
plugin name. When configured with the expected set of plugins
this passes the integration testing suite against 1.3.19.

Adds an abstraction around checking which plugins are installed
to handle the renaming of some plugins between elasticsearch and
opensearch.

Bug: T372769
Change-Id: Ice359e3fa64147e12c65fab19480519f1f5e97e5