deepcat: Don't short-circuit over fetches
When the backend returns more categories than we can support we
were giving a warning to the user and erroring out their query
with no results. This is inconsistent with the rest of cirrus.
Within the query language errors that cancel the query should be
due to the query itself. Something the user can fix and resubmit
their query. In this case there is nothing to fix, the problem is
external to the user.

Update deepcat to no longer short circuit, and update the text of
the warning to make it clear we partially applied the deepcat.

Bug: T376440
Change-Id: If3f69e01a1a2fec8f74a14584819bd2478fdd4fc