Add shard_count to expected indices
Could be interesting to have when reindexing a wiki.
Might possibly save some complexity in the cirrus-reindex-orchestrator
to avoid fetching this info from the backend. But still unsure this is
the right setting to use... the number of source shards might be
slightly more precise to proxy the load a reindex might cause.
We generally increase this number so in the worstcase we would be
overcautious so perhaps acceptable?

Bug: T378382
Change-Id: I600514ae889130cd7b2669b4afd97a853706a9eb