Prefer `self` references wherever possible
Instead of repeating the own class name. I believe this makes the
code easier to read and maintain. For example, it's easier to rename
a class. Or searching via CodeSearch will find the class only once
and not dozens of times. Certainly not a huge difference but still
worth it in my opinion.

Change-Id: Ie29e914acf38d796a924649d01e38400c6d83764