Explicitly Declare icu_normalizer Char Filter
Kuromoji wants the icu_normalizer char filter, instead of the
usual token filter. If ICU is available, copy the token filter
config (including normalization exceptions) to the char filter
config.

(A later stage will remove it for the languages that don't
use it, but for now it's in all the ICU-enabled configs.
A silver lining is that the German config tests that the
exceptions for the token filter get copied to the char
filter.)

Bug: T318269
Change-Id: I7c958641c22ceb7bec0d78d8c26ee13f6a9bc2cb