Mark cirrus api's as internal
All of the API's exposed by cirrussearch are for internal usage,
they generally support either debugging or the implementation of
an external updater process. None are intended for end users, and
none should be used to build extended features on top of.

Public usage of CirrusSearch should always happen through
mediawiki's search api's and not something cirrus specific.

Change-Id: Ida5f7b52980e9b8132bf633597bd865384a6e4f8