sanity: redirect-in-index should skip unindexable link targets
Not all redirects are indexable. Redirects that are to an
external wiki, to a special page, to an invalid title string, and
potentially more options do not represent an indexable redirect.

In those cases skip outputting a link target with the
redirect-in-index problem. The consumer is expected to delete
the page from the index without updating anything else.

Bug: T385571
Change-Id: Id4ba11f8b3f76032fde1a017d3fc6055c0538743