Don't build a document for redirects
Redirects are not directly represented in the search engine, but
document building was happy to go about creating a document that
represented them anyways. Check if given pages are redirects and
don't return a document for them.

In a similar vein, the sanity checker was issuing 'old document'
problems for redirects, but this state doesn't apply to redirects
since they shouldn't exist in the index anyways. Do not issue the
'old document' problem for redirect pages.

Bug: T385571
Change-Id: I730f9b59ebd1c162614a763b6643fca42c1a7e7a