tests: Remove redundant WANObjectCache override
Since 2020 with change Ia60cb0bc18b23, the cache is now enabled in
PHPUnit context by default, so this override is identical to the
default. If it needs to override it, it should use setMainCache()
which takes care of everything else automatically, including
WANObjectCache since that is just a wrapper around the main cache.

Change-Id: I786f6e957b3980f4ec2fdc7124148686cfc222ea