Remove write isolation
Write isolation, along with job queue partitioning which supports
isolation, was added to work around some issues we had in the WMF
deployment related to job queue throughput. But since then the WMF
deployment has migrated away from shipping updates from php and
has a dedicated streaming application to do this work.

We are not aware of any external uses of CirrusSearch that
utilize the multi-cluster wikifarm capabilities of CirrusSearch
and as such it seems most sensible to remove this support and
simplify the operation.

This reverts 5f2ea65b90eb779bc1f647ca88d29d2932298784 and
6ac0d7129634d7453c37f827455a28880b15566b.

Bug: T389429
Change-Id: Ie706ba372972d59ddb4289a05964c3b38f7391f5