Disable writes when running in quibble
In CI quibble jobs we have no ES backend, so writes would fail anyway.
This can be problematic because the jobs get enqueued multiple times and
this may lead to timeouts when processing the job queue under tight time
constraints.

So, add an extension registration callback where we override the default
config when running in quibble, to prevent any writes to be attempted in
the first place.

Bug: T389895
Change-Id: I2dbe568721bb9cb144340c767ea45430c8da3d11