Add a note to cirrus[build]doc output about format stability
As part of the Search action API, CirrusDoc is meant to be used
only internally, for communication between the various Search
components and it isn't a stable interface. To avoid other teams
and/or use cases to start relying on that format, we need to make it
clear in the response itself that this format is unstable.

Bug: T388213
Change-Id: I234cd8054d275c632483db5e029b4ec0271bead9