Gracefully handle BadRevisionException
these exceptions occur when the revision is permanently un-available.
The strategy is to handle these docs and let them being indexed.
Hopefully they'll be hardly findable and won't cause noise in search
results while still being able to be searched for using
hastemplate:CirrusSearchBadRevision in the same vein we handle broken
UTF8 sequences.

Bug: T382904
Change-Id: I0259b76f6c2477ee5652bf25f01a6f78ddf51741