build: drop wdio/codemod
Last version was released 3 years ago, and there have been no
significant changes since then, so this is kind of abandonware. Added in
I1a8440c20dce3cd, presumably to help with the wdio v6 -> v7 migration.
But now we are on v7 upgrading to v8, and codemod does not support any
migrations beyond v6 -> v7. The one thing it still does though, is add
287 packages it depends on. So, drop it from package.json as it is
unused and only causing unneeded packages to be installed, which is
problematic, for security and performance reasons at least.

Change-Id: I765c00fa18790aa732db6aaac29af7ff32926356