regex: Support extended syntax
The search-extra and cirrus-highlighter plugins for opensearch
have been upgraded to support regex anchors (^/$) and short
hand char classes(\d, \s, and \w). Adding support requires
a few changes:

* Fields where we want anchor support have to be indexed
 appropriately. All title fields updated to include regex
 anchor support.

* No changes to the query, char classes are always supported and
 anchor support is auto-detected from the index analyzer.

* While the query doesn't change, the highlighter has to be
 informed about which flavor of regex to highlight. intitle
 and insource queries adjusted to pass the appropriate flavor.

This must not be deployed prior to updating the plugins. Both
index analysis settings and the highlight query require support
from the updated plugin and will otherwise fail.

There will be some inconsistency with intitle queries on initial
deployment, as the highlighter will be using the anchored flavor,
while the query will not be anchored until a reindex has run.
This seems a minor inconvenience, but pretty edge-casey and will
only risk highlighting not matching the results.

Bug: T317599
Change-Id: Ibe8b0d9962ff1d70bba1000e5b8f1152c271d017