Add `suggest_variant` field
We would like to AB test the did-you-mean phrase suggester with
the default title/redirect.title configuration, vs the extended
configuration with title/redirect.title/opening_text. To
facilitate this define a second suggest field which gives us the
opportunity to select the appropriate one with query time
profiles.

This is not intended to be a long term option, once the ab test
has been run this should be reverted and, if desired, the
opening_text flag should be enabled.

Bug: T397083
Change-Id: I6757a1e6b67f4458b30b390ccee56a96e215ae32