UpdateSuggesterIndex: Avoid holding large in-memory batches
When subphrase matching is configured and we are building
suggestions for pages that have a large number of redirects
the ratio of suggest docs to source docs has been observed
to reach 60:1 or higher. Adjust the process to flush the suggest
docs when they reach the indexChunkSize, rather than once we have
process indexChunkSize source documents.

In WMF production we run this script with an indexChunkSize of
3000, which results in ~180k suggest docs being cached in memory
and then shipped to elastic in a single batch.  While this does
work, it leads to undesirable memory usage. On our largest wiki
(by article+redirect count) that has the subphrase completion
suggester enabled this reduced the maximum memory needed to build
the index from ~2GB to ~512MB.

Bug: T395465
Change-Id: I2b138bef07d9dac3f69f0c9a00d7e390a271fea8