CompSuggest: Resolve AUTO fuzziness
We found that AUTO fuzziness was not being applied as expected in
the completion suggester context, it was always resolving to a
constant value of 1. After running an AB test on the WMF wikis
we've found that this looks to generally improve the user
experience. It's not all wins, it's a classic recall/precision
tradeoff, the the recall improvement looks to be larger than
the precision loss by 2x-3x.

As such, always resolve fuzziness which has the end effect of
adding a second character of fuzziness to all autocomplete
queries >= 6 characters in length.

Bug: T397732
Change-Id: I5a6ed997d67b9d31d45b42084b03f321d46c4aa6