Use page_id to sort with search_after
Sorting on the _id will load it in fielddata which resides in mem.
It has not caused problems in the past but better to avoid it.
For context elastic 8 disable loading fielddata on the _id by default,
unsure about opensearch.

While at it also drop the obsolete _type field from the generated dumps.

Change-Id: Ib32e60a342afd25c558ffbf2bbdaa9bc263e1f0e