Try to prevent promoting a bad index in UpdateSuggesterIndex
For several occasions we promoted a bad suggestion index.
The reasons are quite unclear and on this past occurrences we
unfortunately lost the logs of the output of the script.
This patch adds two new checks that are only performed on large indices
(> 100000) and during a rebuild:
- check if the resulting index is not smaller by more than 10% of the
  previous index.
- check if the number of docs in the index is not more than 1% smaller
  than the number of documents generated. Reason why they do not match
  perfectly is still unclear. Could be individual errors in the bulk
  index request or duplicates.

Add extra output with the number of bulk requests sent and various other
stats regarding the outcome of the indexation process.

Bug: T363521
Change-Id: I5844f0f5ee5b46f2087a31974076ff2474ab3932