Use getManagedClusters in the consistency check API
The list of clusters in the cluster API param should be the ones that we
"manage" not all of them.
Determining a default cluster now makes little sense since we can't
really know what cluster is behind a particular php server and it's I
think against the purpose of this API. So better not try to determine a
default one and force the client to be explicit.
This is a breaking a change but this API is marked as internal and there
are no good reasons for not providing a cluster in the params.

Change-Id: I708680850d31908f8bb21dc9c208410b09a724b9