SECURITY: protect cirrusbuilddoc via PoolCounter
This action API property might hit the parser, this may allow an
attacker to starve PHP workers very quickly if used aggressively on
pages that are particularly slow to parse.

This API is meant to be internal and the protection can be bypassed by
setting wgCirrusSearchStreamingUpdaterUsername to a well-known internal
user.

Bug: T401220
Change-Id: I3e8d819868c0491b18368af8e543180e747023c2