UpdateWeighedTags: Support --reset operation
Previously, the maintenance script passed over
$tagPrefix as a string, but WeighedTagsUpdater::resetWeightedTags()
supports only arrays.

Since it does not make any sense to allow multiple occurances
of --tagPrefix in the script, let's construct the array manually
in the script to have something to pass over.

I tested this patch locally and it solves the problem, but
I am not 100% sure whether this is the expected solution
of the problem.

Bug: T403426
Change-Id: Iebcc7bd9f0a6f23d6bb8debffde54534fa8cfe1c