Add profiles for suggest_variant
We added a new `suggest_variant` field that contains the
opening_text, instead of the titles and redirects contained in
the typical `suggest` field. Added profiles are variations of the
three profiles we use in production, but with the phrase
suggester using the variant field. This should be sufficient to
configure an AB test.

While here it was tedious to add duplicates of the various
fallback profiles for each added phrase suggest profiles. Make
this simple by auto-generating the classes of fallback profiles
from a list of supported phrase suggest profiles. This should be
more consistent and less error prone than manually maintaining
the list.

Bug: T390858
Change-Id: I6748d9ffa1312f12b59d1054822d7a5be6fbcb06