Add CirrusSearchAtlernativeIndices
a new config option allowing to build & use alternative indices.
Only completion indices are supported.
The UpdateSuggesterIndex maint script will now inspect this config var
to create multiple completion indices.
At runtime the completion module will attempt to use this index if told
to do so with CirrusSearchCompletionSuggesterUseAltIndexId or via the
cirrusCompletionUseAltIndexId debug param.
The setting is ignored if the alternative index is not explicitly marked
as ready for use.
Adapt cindy to build an alternative index with default_sort enabled.
Flipped the default suggest index to no longer use default sort so that
we can test with cindy that using the alternative index default sort is
being enabled.

Bug: T390268
Depends-On: I4185794d051cc74a8bcdb0cdf26c035f3d901ac4
Change-Id: I299e5edc173529c620ff2e7f3555c31c50a2c5e4