Adjust new dym profiles to match ab test
After analyzing dym results against a sample of queries we've
decided the expensive profile varients aren't providing enough
value to be worth AB testing. The improvements we see come from
either switching to the variant field for better language
statistics, or by changing prefix_length from 2 to 1 which allows
fixing a class of queries that were previously unfixable.

This removes the "expensive" profiles. It additionally tries to
better normalize the profile names to describe what they modify.
As such the new profiles added for ab testing are now:

* default_1 - default profile, with prefix_length=1
* default_1v - default profile, with prefix_length=1 and
  field=suggst_variant

Bug: T390858
Change-Id: I8ad22367bedb443b46b0602ef69508748348df3c