Implement syntax for range queries on dates
Expands the set of search keywords offered to include range
queries against the article creation and last edit dates.
Implemented as a generic keyword that can attach to any date
field.

All date comparisons are rounded to the same precision as the
units provided. This means <2024 is < 2024-01-01, and >2024 is
>= 2025-01-01. Exact dates can be provided for year, year/month,
and year/month/day.  Additionally relative queries can specify
an offset from now in hours, days, months, or years.

Bug: T403593
Change-Id: I968397e9e735134108b05685a5919095f0f0c3a7