tests: skip EventBus tests when it is not loaded
On CI, we would like to remove the recursive injection of dependencies.
CirrusSearch has integration tests with EventBus which is otherwise not
needed by extensions depending on CirrusSearch.

Skip the EventBus integration tests when it is not loaded. This will
save us from having EventBus added on CI for any extension loading
CirrusSearch.

Bug: T389998
Change-Id: I45ffd967f3f3a246d65e9072c3f11ee142339870