Stop using browser_tests profile in test fixtures
Upcoming functionality will expose the full text query profiles
to users via the mediawiki apis. As such we want to remove the
browser_tests profile from the set of available profiles as it
was never intended for public usage.

As a first step stop using the profile in our test suite. A
follow-up patch will move the profile from the default conf
to tests/jenkins/FullyFeaturedConfig.php, which will mean the
profile will no longer be available during phpunit tests.

Change-Id: Iba578d5db578a48401bc1fab84538b666d8f7aab