Refactor SecondTrySearch
refactor SecondTrySearch as a separate class per strategy.
Also change the signature such that if nothing is changed no extra
candidates are required to be tested.

Unfortunately for this I had to slightly change the logic since we might
lower-case some strings in the middle of the algorithm.

Returning an array might also be interesting for language converter
variants that might return multiple candidates.

Bug: T402864
Change-Id: If062639627c2da951d583623b38960acfeb62b7c