Revert "Limit suggest_variant to only opening_text"
This reverts commit 4d049e531b412044083a26caf52192c76eb7eecb.

We've run the AB test against suggest_variant and were surprised
to find that for most requests the existing suggest field
outperforms it, even though in our testing we saw several types
of queries that were only fixable with the variant field.

One plausible answer is that the title and redirect.title content
provides valuable statistics, but we discarded them. As such we
will re-run the test with the title and redirec.title included in
suggest_variant field.

Bug: T407432
Change-Id: Ib60b459d48ff06afdcc2540b5c1ba25c1806796b