Prune weighted_tags deleted marker during reindex
We have managed to index a bunch of these, and while it's not
causing direct problems they should still be filtered out of the
live index. These are a little annoying to delete directly, if we
ask the super_detect_noop script to delete them it will see that
the update and the current content already match and it's noop's
the change. We could issue two sequential changes, first to
inject a fake tag and then a second to delete it, but that seems
error prone and against enough docs plausible to run into a race
condition.

Resolve the problem by adding a script that is run during
reindexing that will prune the markers. This will, at a minimum,
allow us to verify that we are no longer shipping updates that
cause these to be added to the index instead of interpreted as
delete markers.

Bug: T406205
Change-Id: I00a7ab186bfbb2aad6b05d4675dc187fa34176e9