ForceSearchIndex: fail fast when updates disabled
When updates are disabled we would still initialize and run the
jobs, only for the jobs to not do anything because they were
disabled. While this is functionaly correct, it isn't the most
obvious way to operate. Instead check before doing anything and
fail with an appropriate message if updates are disabled.

Bug: T408678
Change-Id: I9bb2f97a97a47253cedd6c99129d83fa4b2e0069