regex: Disable anchored regex support on elasticsearch
Anchored regex's are only supported in the newer version of the
plugin on opensearch, the elasticsearch plugin is older and does
not have that feature. As such disable the support at index time.
There are no query-time related changes, updating the mapping
is sufficient.

Implemented by sourcing version information from the server banner
and passing it down to the AnalysisConfigBuilder instances. This
is always passed with plugins, could plausibly unify the plugins
and version info into some sort of RemoteServerDescription, but
that is left for the future.

Bug: T409070
Change-Id: I0ca44cffbdc6184c1b481c0a30b716e5be82626f