Revert "regex: Disable anchored regex support on elasticsearch"
This reverts commit 1ffe5425b9ef6be9b6add44a785582a4ec88221d.

On further testing this was found to be inadequate. At query time
we vary the regex flavor asked of the highlighter, but we don't
want to round trip to the server to find out what version it is
on every request. We could add a config flag, but that seems
error prone.

Instead a follow patch will remove support for anchored regexes
in 1.45. That feature will have to be delayed to 1.46 when we
remove support for elasticsearch.

Bug: T409070
Change-Id: I9b40b248937456af10c9e377136addfd3b5e11ea