Streamline weight calculations in NamespacesFunctionScoreBuilder
I found this loop surprisingly confusing when all it really is
is a map from a weight value to a list of namespaces. The only
disadvantage of my proposal is that the weight is converted to a
string when it becomes an array key and back to a float. But I
think this is acceptable. There is no data loss and the performance
loss is effectively zero.

Change-Id: Ibafd1ba1c6aa25ef7948c6dffe2a971894520fa9